SigilHive/http/requirements.txt
aiohttp
langchain
langchain-google-genai
python-dotenv
uvloop
aiofiles
confluent_kafka